languages: Move to neutral Spanish (es_ES -> es)
As with Spanish translations of GNOME apps, neutral Spanish is normally
used so that contributions are shared among the different variants of
Spanish.

Skimming through `keypunch/data/word_lists/es_ES.txt`, I could not find any
specific words for any of the variants, so we are good on that side too.