data/meson.build: Add gsed as sed substitute
BSD systems don't use GNU sed by default, but may have `gsed` installed.
The command used doesn't work with the default sed, so now we'll prefer
it if running on a BSD system or if it's available at all (because there
are other cases where this might happen, like maybe Alpine Linux).

I tested on FreeBSD, but not on other BSD's.